Add configurable date format.
This pulls in moment.js, but not through the normal addon in an attempt
to keep filesize down a bit (we dont need all the extra components or
the timezone support).

For now this is configured via ENV, but I'd like to do a followup PR to
move config into a separate file (see #17).

- adds default dateFormat (defaults to MM-DD-YYYY)
- adds article-info component
- adds format-date helper

Fixes #2.